+++ DataAutoFramework/TestCases/TestPageContent.cs
@@ -1,8 +1,7 @@
 using NUnit.Framework.Legacy;
 using NUnit.Framework;
-using Microsoft.Playwright;
 using System.Text.Json;
-using System.Text.RegularExpressions;
+using DataAutoFramework.Utilities;
 
 namespace DataAutoFramework.TestCases
 {
@@ -29,27 +28,9 @@ namespace DataAutoFramework.TestCases
         public async Task TestDuplicateServiceByContent(string testLink)
         {
 
-            using var playwright = await Playwright.CreateAsync();
-            var browser = await playwright.Chromium.LaunchAsync(new BrowserTypeLaunchOptions { Headless = true });
-            var page = await browser.NewPageAsync();
-
-            await page.GotoAsync(testLink);
-
-            HashSet<string> set = new HashSet<string>();
             List<string> duplicateTexts = new List<string>();
-
-            var aElements = await page.Locator("li.has-three-text-columns-list-items.is-unstyled a[data-linktype='relative-path']").AllAsync();
-
-            foreach (var aElement in aElements)
-            {
-                var textContent = await aElement.InnerTextAsync();
-                if (!set.Add(textContent)) 
-                {
-                    duplicateTexts.Add(textContent);
-                }
-            }
-
-            await browser.CloseAsync();
+            var checkPageContent = new CheckPageContent();
+            await checkPageContent.CheckDuplicateServiceByContent(testLink, duplicateTexts);
 
             ClassicAssert.Zero(duplicateTexts.Count, testLink + " has duplicate service at " + string.Join(",", duplicateTexts));
         }
@@ -58,33 +39,10 @@ namespace DataAutoFramework.TestCases
         [TestCaseSource(nameof(SiderTestLinks))]
         public async Task TestDuplicateServiceBySider(string testLink)
         {
-            using var playwright = await Playwright.CreateAsync();
-            var browser = await playwright.Chromium.LaunchAsync(new BrowserTypeLaunchOptions { Headless = true });
-            var page = await browser.NewPageAsync();
-
-            await page.GotoAsync(testLink);
-            await page.WaitForSelectorAsync("li.border-top.tree-item.is-expanded");
-
-            var parentLi = await page.QuerySelectorAsync("li.border-top.tree-item.is-expanded");
-
-            var liElements = await parentLi.QuerySelectorAllAsync("ul.tree-group > li[aria-level='2']");
-
-            HashSet<string> set = new HashSet<string>();
+            
             List<string> duplicateList = new List<string>();
-
-            foreach (var element in liElements)
-            {
-                var text = await element.InnerTextAsync();
-                if (text != "Overview")
-                {
-                    if (!set.Add(text))
-                    {
-                        duplicateList.Add(text);
-                    }
-                }
-            }
-
-            await browser.CloseAsync();
+            var checkPageContent = new CheckPageContent();
+            await checkPageContent.CheckDuplicateServiceBySider(testLink, duplicateList);
 
             ClassicAssert.Zero(duplicateList.Count, testLink + " has duplicate service at " + string.Join(",", duplicateList));
         }
@@ -94,23 +52,9 @@ namespace DataAutoFramework.TestCases
         public async Task TestGarbledText(string testLink)
         {
             var errorList = new List<string>();
-            var playwright = await Playwright.CreateAsync();
-            var browser = await playwright.Chromium.LaunchAsync(new BrowserTypeLaunchOptions { Headless = true });
-            var page = await browser.NewPageAsync();
-            await page.GotoAsync(testLink);
-            var pLocators = await page.Locator("p").AllAsync();
+            var checkPageContent = new CheckPageContent();
+            await checkPageContent.CheckGarbledText(testLink, errorList);
 
-            foreach (var pLocator in pLocators)
-            {
-                var text = await pLocator.TextContentAsync();
-
-                if (Regex.IsMatch(text, @":[\w]+(?:\s+[\w]+){0,2}:"))
-                {
-                    errorList.Add(text);
-                }
-            }
-            
-            await browser.CloseAsync();
             ClassicAssert.Zero(errorList.Count, testLink + " has garbled text" + string.Join(",", errorList));
         }
 
@@ -118,30 +62,9 @@ namespace DataAutoFramework.TestCases
         [TestCaseSource(nameof(TestLinks))]
         public async Task TestIsTableEmpty(string testLink)
         {
-            var count = 0;
-            var playwright = await Playwright.CreateAsync();
-            var browser = await playwright.Chromium.LaunchAsync(new BrowserTypeLaunchOptions { Headless = true });
-            var page = await browser.NewPageAsync();
-            await page.GotoAsync(testLink);
+            var checkPageTable = new CheckPageContent();
+            var count = await checkPageTable.CheckIsTableEmpty(testLink);
 
-            var tableLocator = page.Locator("table");
-            //var tableLocator = page.Locator("table:not([aria-label*='Package'])");
-            var rows = await tableLocator.Locator("tr").AllAsync();
-
-            foreach (var row in rows)
-            {
-                var cells = await row.Locator("td, th").AllAsync();
-                foreach (var cell in cells)
-                {
-                    var textContent = await cell.TextContentAsync();
-                    if (string.IsNullOrWhiteSpace(textContent))
-                    {
-                        count++;
-                    } 
-                }
-            }
-            
-            await browser.CloseAsync();
             ClassicAssert.Zero(count, testLink + " has table is empty.");
         }
     }
